+++ test/lms_test.c
@@ -7,6 +7,12 @@
  * https://www.openssl.org/source/license.html
  */
 
+/*
+ * Rather than creating a nearly identical file for HSS,
+ * this file is used to test either LMS or HSS public key and signature
+ * verification. It shares the code since the public key encoding only differs
+ * by an additional 4 bytes (for L).
+ */
 #include <openssl/core_names.h>
 #include <openssl/decoder.h>
 #include <openssl/evp.h>
@@ -19,6 +25,7 @@ typedef enum OPTION_choice {
     OPT_ERR = -1,
     OPT_EOF = 0,
     OPT_CONFIG_FILE,
+    OPT_KEY_ALG,
     OPT_TEST_ENUM
 } OPTION_CHOICE;
 
@@ -26,8 +33,12 @@ static OSSL_LIB_CTX *libctx = NULL;
 static char *propq = NULL;
 static OSSL_PROVIDER *nullprov = NULL;
 static OSSL_PROVIDER *libprov = NULL;
+/* Chose LMS by default */
+static char *key_alg = "LMS";
+static LMS_ACVP_TEST_DATA *testdata = lms_testdata;
+static size_t testdata_len = OSSL_NELEM(lms_testdata);
 
-static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
+static EVP_PKEY *pubkey_from_data(const unsigned char *data, size_t datalen)
 {
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
@@ -37,7 +48,7 @@ static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
     params[0] = OSSL_PARAM_construct_octet_string(OSSL_PKEY_PARAM_PUB_KEY,
                                                   (unsigned char *)data, datalen);
     params[1] = OSSL_PARAM_construct_end();
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", propq))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, propq))
         && TEST_int_eq(EVP_PKEY_fromdata_init(ctx), 1)
         && (EVP_PKEY_fromdata(ctx, &key, EVP_PKEY_PUBLIC_KEY, params) == 1);
     if (ret == 0) {
@@ -51,7 +62,7 @@ static EVP_PKEY *lms_pubkey_from_data(const unsigned char *data, size_t datalen)
 static int lms_bad_pub_len_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     size_t publen = 0;
     unsigned char pubdata[128];
@@ -65,7 +76,7 @@ static int lms_bad_pub_len_test(void)
     for (publen = 0; publen <= td->publen + 16; publen += 3) {
         if (publen == td->publen)
             continue;
-        if (!TEST_ptr_null(pkey = lms_pubkey_from_data(pubdata, publen)))
+        if (!TEST_ptr_null(pkey = pubkey_from_data(pubdata, publen)))
             goto end;
     }
     ret = 1;
@@ -78,51 +89,82 @@ end:
     return ret == 1;
 }
 
-static int lms_pubkey_decoder_fail_test(void)
+/* params[] contains a list of invalid public key data */
+static int do_pubkey_decoder_fail(const OSSL_PARAM params[])
 {
     int ret = 0;
     EVP_PKEY *pkey = NULL;
     OSSL_DECODER_CTX *dctx = NULL;
     int selection = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
     const unsigned char *pdata;
     size_t pdatalen;
-    static const unsigned char pub_bad_LMSType[] = {
-        0x00, 0x00, 0x00, 0xAA
-    };
-    static const unsigned char pub_bad_OTSType[] = {
-        0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
-    };
+    const OSSL_PARAM *p = params;
 
-    if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pkey, NULL, NULL, "LMS",
+    if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pkey, NULL, NULL, key_alg,
                                                        selection,
                                                        libctx, NULL)))
         return 0;
 
-    pdata = td->pub;
-    pdatalen = 3;
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
-    pdatalen = SIZE_MAX;
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
-    pdata = pub_bad_LMSType;
-    pdatalen = sizeof(pub_bad_LMSType);
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
-
-    pdata = pub_bad_OTSType;
-    pdatalen = sizeof(pub_bad_OTSType);
-    if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
-        goto end;
+    while (p->key != NULL) {
+        pdata = p->data;
+        pdatalen = p->data_size;
+        if (!TEST_false(OSSL_DECODER_from_data(dctx, &pdata, &pdatalen)))
+            goto end;
+        p++;
+    }
 
     ret = 1;
 end:
     EVP_PKEY_free(pkey);
     OSSL_DECODER_CTX_free(dctx);
     return ret;
+}
+
+static int lms_pubkey_decoder_fail_test(void)
+{
+    OSSL_PARAM params[6], *p = params;
+
+    *p++ = OSSL_PARAM_construct_octet_string("Invalid Small Pub Len",
+                                             (void *)testdata->pub, 3);
+    *p++ = OSSL_PARAM_construct_octet_string("Invalid Pub Len",
+                                             (void *)testdata->pub, SIZE_MAX);
+    if (testdata == lms_testdata) {
+        static const unsigned char lms_pub_bad_LMSType[] = {
+            0x00, 0x00, 0x00, 0xAA
+        };
+        static const unsigned char lms_pub_bad_OTSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+        *p++ = OSSL_PARAM_construct_octet_string("Bad LMS type",
+                                                 (void *)lms_pub_bad_LMSType,
+                                                 sizeof(lms_pub_bad_LMSType));
+        *p++ = OSSL_PARAM_construct_octet_string("Bad OTS type",
+                                                 (void *)lms_pub_bad_OTSType,
+                                                 sizeof(lms_pub_bad_OTSType));
+    } else {
+        static const unsigned char hss_pub_bad_L[] = {
+            0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x01
+        };
+        static const unsigned char hss_pub_bad_LMSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+        static const unsigned char hss_pub_bad_OTSType[] = {
+            0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0xAA
+        };
+
+        *p++ = OSSL_PARAM_construct_octet_string("bad L",
+                                                 (void *)hss_pub_bad_L,
+                                                 sizeof(hss_pub_bad_L));
+        *p++ = OSSL_PARAM_construct_octet_string("bad LMS type",
+                                                 (void *)hss_pub_bad_LMSType,
+                                                 sizeof(hss_pub_bad_LMSType));
+        *p++ = OSSL_PARAM_construct_octet_string("bad OTS type",
+                                                 (void *)hss_pub_bad_OTSType,
+                                                 sizeof(hss_pub_bad_OTSType));
+    }
+    *p = OSSL_PARAM_construct_end();
+
+    return do_pubkey_decoder_fail(params);
 }
 
 static EVP_PKEY *key_decode_from_bio(BIO *bio, const char *keytype)
@@ -161,7 +203,7 @@ static EVP_PKEY *key_decode_from_data(const unsigned char *data, size_t datalen,
 static int lms_key_decode_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td1 = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td1 = &testdata[0];
     EVP_PKEY *key = NULL;
 
     ret = TEST_ptr(key = key_decode_from_data(td1->pub, td1->publen, NULL));
@@ -172,14 +214,14 @@ static int lms_key_decode_test(void)
 static int lms_pubkey_decoder_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY *pub = NULL;
     OSSL_DECODER_CTX *dctx = NULL;
     const unsigned char *data;
     size_t data_len;
 
     if (!TEST_ptr(dctx = OSSL_DECODER_CTX_new_for_pkey(&pub, "xdr", NULL,
-                                                       "LMS",
+                                                       key_alg,
                                                        OSSL_KEYMGMT_SELECT_PUBLIC_KEY,
                                                        libctx, NULL)))
         goto err;
@@ -198,8 +240,8 @@ static int lms_key_eq_test(void)
 {
     int ret = 0;
     EVP_PKEY *key[4];
-    LMS_ACVP_TEST_DATA *td1 = &lms_testdata[0];
-    LMS_ACVP_TEST_DATA *td2 = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td1 = &testdata[0];
+    LMS_ACVP_TEST_DATA *td2 = &testdata[1];
     size_t i;
 #ifndef OPENSSL_NO_EC
     EVP_PKEY *eckey = NULL;
@@ -208,9 +250,9 @@ static int lms_key_eq_test(void)
     for (i = 0; i < OSSL_NELEM(key); i++)
         key[i] = NULL;
 
-    if (!TEST_ptr(key[0] = lms_pubkey_from_data(td1->pub, td1->publen))
-            || !TEST_ptr(key[1] = lms_pubkey_from_data(td1->pub, td1->publen))
-            || !TEST_ptr(key[2] = lms_pubkey_from_data(td2->pub, td2->publen))
+    if (!TEST_ptr(key[0] = pubkey_from_data(td1->pub, td1->publen))
+            || !TEST_ptr(key[1] = pubkey_from_data(td1->pub, td1->publen))
+            || !TEST_ptr(key[2] = pubkey_from_data(td2->pub, td2->publen))
             || !TEST_ptr(key[3] = key_decode_from_data(td1->pub, td1->publen,
                                                        NULL)))
         goto end;
@@ -238,11 +280,11 @@ end:
 static int lms_key_validate_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY_CTX *vctx = NULL;
     EVP_PKEY *key = NULL;
 
-    if (!TEST_ptr(key = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(key = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_PKEY_CTX_new_from_pkey(libctx, key, NULL)))
         goto end;
@@ -256,13 +298,13 @@ end:
 static int lms_verify_test(int tst)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[tst];
+    LMS_ACVP_TEST_DATA *td = &testdata[tst];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
 
-    ret = TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    ret = TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         && TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL))
         && TEST_int_eq(EVP_PKEY_verify_message_init(ctx, sig, NULL), 1)
         && TEST_int_eq(EVP_PKEY_verify(ctx, td->sig, td->siglen,
@@ -277,17 +319,17 @@ static int lms_verify_test(int tst)
 static int lms_digest_verify_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY *pub = NULL;
     EVP_MD_CTX *vctx = NULL;
 
-    if (!TEST_ptr(pub = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(pub = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_MD_CTX_new()))
         goto err;
     /* Only one shot mode is supported, streaming fails to initialise */
     if (!TEST_int_eq(EVP_DigestVerifyInit_ex(vctx, NULL, NULL, libctx, NULL,
-                                             pub, NULL), 0))
+                                             pub, NULL), testdata == hss_testdata))
         goto err;
     ret = 1;
  err:
@@ -299,11 +341,11 @@ static int lms_digest_verify_fail_test(void)
 static int lms_digest_signing_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY *pub = NULL;
     EVP_MD_CTX *vctx = NULL;
 
-    if (!TEST_ptr(pub = lms_pubkey_from_data(td->pub, td->publen)))
+    if (!TEST_ptr(pub = pubkey_from_data(td->pub, td->publen)))
         return 0;
     if (!TEST_ptr(vctx = EVP_MD_CTX_new()))
         goto err;
@@ -320,13 +362,13 @@ static int lms_digest_signing_fail_test(void)
 static int lms_message_signing_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
 
-    ret = TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    ret = TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+        && TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
         && TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL))
         && TEST_int_eq(EVP_PKEY_sign_message_init(ctx, sig, NULL), -2);
 
@@ -341,7 +383,7 @@ static int lms_paramgen_fail_test(void)
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
 
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", NULL))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, NULL))
         && TEST_int_eq(EVP_PKEY_paramgen_init(ctx), -2);
 
     EVP_PKEY_CTX_free(ctx);
@@ -353,7 +395,7 @@ static int lms_keygen_fail_test(void)
     int ret;
     EVP_PKEY_CTX *ctx = NULL;
 
-    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, "LMS", NULL))
+    ret = TEST_ptr(ctx = EVP_PKEY_CTX_new_from_name(libctx, key_alg, NULL))
         && TEST_int_eq(EVP_PKEY_keygen_init(ctx), -2);
 
     EVP_PKEY_CTX_free(ctx);
@@ -363,11 +405,11 @@ static int lms_keygen_fail_test(void)
 static int lms_verify_fail_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[0];
+    LMS_ACVP_TEST_DATA *td = &testdata[0];
     EVP_PKEY_CTX *ctx = NULL;
     EVP_PKEY *pkey = NULL;
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
             || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
     /* Only one shot mode is supported, streaming fails to initialise */
@@ -383,7 +425,7 @@ end:
 static int lms_verify_bad_sig_test(void)
 {
     int ret = 0, i = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
     EVP_PKEY_CTX *ctx = NULL;
@@ -404,8 +446,8 @@ static int lms_verify_bad_sig_test(void)
     if (sig_data == NULL)
         return 0;
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-            || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+            || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
             || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
 
@@ -444,7 +486,7 @@ end:
 static int lms_verify_bad_sig_len_test(void)
 {
     int ret = 0;
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
     EVP_PKEY_CTX *ctx = NULL;
@@ -452,8 +494,8 @@ static int lms_verify_bad_sig_len_test(void)
     const int step = 3;
     unsigned char sigdata[4096];
 
-    if (!TEST_ptr(pkey = lms_pubkey_from_data(td->pub, td->publen))
-            || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL))
+    if (!TEST_ptr(pkey = pubkey_from_data(td->pub, td->publen))
+            || !TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL))
             || !TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
         goto end;
 
@@ -488,7 +530,7 @@ end:
 
 static int lms_verify_bad_pub_sig_test(void)
 {
-    LMS_ACVP_TEST_DATA *td = &lms_testdata[1];
+    LMS_ACVP_TEST_DATA *td = &testdata[1];
     int ret = 0, i = 0;
     EVP_PKEY *pkey = NULL;
     EVP_SIGNATURE *sig = NULL;
@@ -500,13 +542,13 @@ static int lms_verify_bad_pub_sig_test(void)
     if (!TEST_ptr(pub = OPENSSL_memdup(td->pub, td->publen)))
         return 0;
 
-    if (!TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, "LMS", NULL)))
+    if (!TEST_ptr(sig = EVP_SIGNATURE_fetch(libctx, key_alg, NULL)))
         goto end;
 
     for (i = 0; i < (int)td->publen; i += step) {
         pub[i] ^= 1; /* corrupt a byte */
         /* Corrupting the public key may cause the key load to fail */
-        pkey = lms_pubkey_from_data(pub, td->publen);
+        pkey = pubkey_from_data(pub, td->publen);
         if (pkey != NULL) {
             if (!TEST_ptr(ctx = EVP_PKEY_CTX_new_from_pkey(libctx, pkey, NULL)))
                 goto end;
@@ -542,6 +584,7 @@ const OPTIONS *test_get_options(void)
         OPT_TEST_OPTIONS_DEFAULT_USAGE,
         { "config", OPT_CONFIG_FILE, '<',
           "The configuration file to use for the libctx" },
+        { "keytype", OPT_KEY_ALG, 's', "set to either LMS or HSS" },
         { NULL }
     };
     return options;
@@ -563,6 +606,9 @@ int setup_tests(void)
             break;
         case OPT_TEST_CASES:
             break;
+        case OPT_KEY_ALG:
+            key_alg = opt_arg();
+            break;
         default:
         case OPT_ERR:
             return 0;
@@ -570,6 +616,10 @@ int setup_tests(void)
     }
     if (!test_get_libctx(&libctx, &nullprov, config_file, &libprov, NULL))
         return 0;
+    if (strcmp(key_alg, "HSS") == 0) {
+        testdata = hss_testdata;
+        testdata_len = OSSL_NELEM(hss_testdata);
+    }
 
     ADD_TEST(lms_bad_pub_len_test);
     ADD_TEST(lms_key_validate_test);
@@ -577,7 +627,7 @@ int setup_tests(void)
     ADD_TEST(lms_key_decode_test);
     ADD_TEST(lms_pubkey_decoder_test);
     ADD_TEST(lms_pubkey_decoder_fail_test);
-    ADD_ALL_TESTS(lms_verify_test, OSSL_NELEM(lms_testdata));
+    ADD_ALL_TESTS(lms_verify_test, testdata_len);
     ADD_TEST(lms_verify_fail_test);
     ADD_TEST(lms_digest_verify_fail_test);
     ADD_TEST(lms_digest_signing_fail_test);
@@ -587,7 +637,6 @@ int setup_tests(void)
     ADD_TEST(lms_verify_bad_sig_test);
     ADD_TEST(lms_verify_bad_sig_len_test);
     ADD_TEST(lms_verify_bad_pub_sig_test);
-
     return 1;
 }